Move class used as query parameters to Models folder
