+++ AniHubApp/AniHubApp/Models/AnimeListQueryParams.cs
@@ -1,0 +1,13 @@
+﻿using Refit;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace AniHubApp.Models
+{
+    public class AnimeListQueryParams
+    {
+        [AliasAs("season")]
+        public string SeasonPeriod { get; set; }
+    }
+}
