+++ AniHubApp/AniHubApp/Models/AnimeSongsListQueryParams.cs
@@ -1,0 +1,13 @@
+﻿using Refit;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace AniHubApp.Models
+{
+    public class AnimeSongsListQueryParams
+    {
+        [AliasAs("anime_id")]
+        public string AnimeId { get; set; }
+    }
+}
