+++ AniHubApp/AniHubApp/Services/IAniApiService.cs
@@ -4,14 +4,15 @@ using System.Collections.Generic;
 using System.Text;
 using System.Threading.Tasks;
 using static AniHubApp.Configuration;
+using AniHubApp.Models;
 
 namespace AniHubApp.Services
 {
     public interface IAniApiService
     {
-        Task<AnimeListResponse> GetAnimesAsync(AniAPI.AnimeListQueryParams queryParams = null);
+        Task<AnimeListResponse> GetAnimesAsync(AnimeListQueryParams queryParams = null);
         Task<AnimeResponse> GetAnimeByIdAsync(int animeId);
-        Task<SongListResponse> GetAnimeSongsAsync(AniAPI.AnimeSongsListQueryParams queryParams = null);
+        Task<SongListResponse> GetAnimeSongsAsync(AnimeSongsListQueryParams queryParams = null);
         Task<SongResponse> GetAnimeSongByIdAsync(int animeSongId);
     }
 }
